setup_gitconfig.sh works with main script

+++ script/setup_scripts/lib/file.bash
@@ -29,12 +29,34 @@ export_files()
     shift 2
     local array_files=("$@")
 
-    echo "Copying files from '$source_path/' to '$dest_path/'..."
+    # Check if source files exists
     for file in "${array_files[@]}"
     do
         [[ -f "$source_path/$file" ]]
         eval_cmd "Necessary file does not exist:\n    $source_path/$file"
+    done
 
+    # Check if all existing dest files equal source files
+    local files_already_exists=true
+    for file in "${array_files[@]}"
+    do
+        if ! [[ -f "$dest_path/$file" ]] || \
+           ! cmp "$source_path/$file" "$dest_path/$file"
+        then
+            files_already_exists=false
+            break
+        fi
+    done
+
+    if [[ "$files_already_exists" == "true" ]]
+    then
+        return_value_export_files='already done'
+        return 0
+    fi
+
+    echo "Copying files from '$source_path/' to '$dest_path/'..."
+    for file in "${array_files[@]}"
+    do
         cp "$source_path/$file" "$dest_path/$file"
         eval_cmd "Could not copy file:\n    $source_path/$file\nto\n    $dest_path/$file"
